功能按钮大小调整

+++ Itsycal/other/NSBezierPath+quartz.m
@@ -12,7 +12,7 @@
 // This method works only in OS X v10.2 and later.
 - (CGPathRef)quartzPath
 {
-    int i, numElements;
+    NSInteger i, numElements;
  
     // Need to begin a path here.
     CGPathRef           immutablePath = NULL;
